Refactor: 패키지 명 PlannerDetail 에서 Planner로 변경
application 에서 createPlannerDetail 테스트 코드 변경
plannerDetailServiceTest 에서 검증하던 createPlannerDetail 을 application으로 변경
