+++ src/main/java/com/zero/triptalk/place/entity/Place.java
@@ -1,15 +1,12 @@
 package com.zero.triptalk.place.entity;
 
 
-import com.zero.triptalk.plannerdetail.entity.PlannerDetail;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
 
 @Getter
 @Entity
