+++ src/main/java/com/zero/triptalk/planner/dto/PlannerDetailDto.java
@@ -1,7 +1,7 @@
-package com.zero.triptalk.plannerdetail.dto;
+package com.zero.triptalk.planner.dto;
 
 import com.zero.triptalk.place.entity.Place;
-import com.zero.triptalk.plannerdetail.entity.PlannerDetail;
+import com.zero.triptalk.planner.entity.PlannerDetail;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
