+++ src/main/java/com/zero/triptalk/planner/dto/PlannerRequest.java
@@ -1,7 +1,7 @@
-package com.zero.triptalk.plannerdetail.dto;
+package com.zero.triptalk.planner.dto;
 
 
-import com.zero.triptalk.plannerdetail.entity.Planner;
+import com.zero.triptalk.planner.entity.Planner;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
